assign tag to product frontend , adding profile picture in employee list screen
